test: update simulator
